SCB-856 Refactor TCC alpha server package.
